refactor(slack): split bridge into bounded modules
Preserve the Experiment Runner Slack Socket Mode facade and CLI while moving config, parsing, rendering, audit/rate/idempotency, dispatch/live approval, and optional Slack transport into bounded modules.

Record mandatory pre-open role execution, premortem dispositions, PR #1852 closeout, and the separate root-artifact hygiene follow-up in governance artifacts.
